+++ seeds/initial-seed.js
@@ -6,7 +6,7 @@ const passTypeData = require('./data/pass_type');
 const daypassData = require('./data/daypass');
 const orderData = require('./data/order');
 const orderTypeData = require('./data/order_type');
-const chargeTypeData = require('./data/charge_type');
+const transactionTypeData = require('./data/transaction_type');
 const gymAmenitiesData = require('./data/gym_amenities');
 const blackoutDatesData = require('./data/blackout_date');
 const gymHoursData = require('./data/gym_hours');
@@ -17,8 +17,8 @@ const visitData = require('./data/visit');
 exports.seed = function (knex, Promise) {
   return knex('order').del()
     .then(() => knex('customer').del())
-    .then(() => knex('charge').del())
-    .then(() => knex('charge_type').del())
+    .then(() => knex('transaction').del())
+    .then(() => knex('transaction_type').del())
     .then(() => knex('visit').del())
     .then(() => knex('daypass').del())
     .then(() => knex('pass_type').del())
@@ -49,8 +49,8 @@ exports.seed = function (knex, Promise) {
     .then(() => knex.raw("SELECT setval('blackout_date_id_seq', (SELECT COUNT(*) FROM blackout_date))"))
     .then(() => knex('pass_type').insert(passTypeData))
     .then(() => knex.raw("SELECT setval('pass_type_id_seq', (SELECT COUNT(*) FROM pass_type))"))
-    .then(() => knex('charge_type').insert(chargeTypeData))
-    .then(() => knex.raw("SELECT setval('charge_type_id_seq', (SELECT COUNT(*) FROM charge_type))"))
+    .then(() => knex('transaction_type').insert(transactionTypeData))
+    .then(() => knex.raw("SELECT setval('transaction_type_id_seq', (SELECT COUNT(*) FROM transaction_type))"))
     .then(() => knex('listing').insert(listingData))
     .then(() => knex.raw("SELECT setval('listing_id_seq', (SELECT COUNT(*) FROM listing))"))
     .then(() => knex('order_type').insert(orderTypeData))
